Minor changes : set azimuth aka float value (this digital filter)

+++ Inc/main.h
@@ -89,7 +89,7 @@ enum {
 #endif
 
 typedef struct compas_data_t {
-	uint16_t angleHMC;
+	float angleHMC;
 	float tempHMC;
 } compas_data_t;
